requirements.txt
Flask==0.10.1
python-dotenv==0.1.2
Flask-RESTful==0.3.4
SQLAlchemy==1.0.9
Flask-SQLAlchemy==2.0
psycopg2==2.6.1
librabbitmq==1.6.1
celery==3.1.19
amqp==1.4.7
Flask-Celery==2.4.3
Flask-Migrate==1.6.0
coverage==4.0.2
pandas==0.17.0
numpy==1.10.1
redis==2.10.5
Jinja2==2.8
uWSGI==2.0.11.2
gevent==1.1b6
gunicorn==19.3.0
Flask-Script==2.0.5
Flask-Alembic==1.2.1
Flask-Admin==1.3.0
Flask-Dogpile-Cache==0.2
Flask-UUID==0.2
sqlalchemy-migrate==0.7.2
Flask-Testing
